+++ realm/realm-library/src/androidTest/java/io/realm/BulkInsertTests.java
@@ -68,11 +68,10 @@ public class BulkInsertTests {
     public final TestRealmConfigurationFactory configFactory = new TestRealmConfigurationFactory();
 
     private Realm realm;
-    private RealmConfiguration realmConfig;
 
     @Before
     public void setUp() {
-        realmConfig = configFactory.createConfiguration();
+        RealmConfiguration realmConfig = configFactory.createConfiguration();
         realm = Realm.getInstance(realmConfig);
     }
 
@@ -97,9 +96,9 @@ public class BulkInsertTests {
 
         AllJavaTypes allTypes = new AllJavaTypes();
         allTypes.setFieldString("String");
-        allTypes.setFieldLong(1l);
-        allTypes.setFieldFloat(1f);
-        allTypes.setFieldDouble(1d);
+        allTypes.setFieldLong(1L);
+        allTypes.setFieldFloat(1F);
+        allTypes.setFieldDouble(1D);
         allTypes.setFieldBoolean(true);
         allTypes.setFieldDate(date);
         allTypes.setFieldBinary(new byte[]{1, 2, 3});
@@ -112,6 +111,7 @@ public class BulkInsertTests {
 
         AllJavaTypes realmTypes = realm.where(AllJavaTypes.class).findFirst();
 
+        assertNotNull(realmTypes);
         assertNotSame(allTypes, realmTypes); // Objects should not be considered equal
         assertEquals(allTypes.getFieldString(), realmTypes.getFieldString()); // But they contain the same data
         assertEquals(allTypes.getFieldLong(), realmTypes.getFieldLong());
@@ -148,6 +148,7 @@ public class BulkInsertTests {
         realm.commitTransaction();
 
         AllTypesRealmModel first = realm.where(AllTypesRealmModel.class).findFirst();
+        assertNotNull(first);
         assertEquals(allTypes.columnString, first.columnString);
         assertEquals(allTypes.columnLong, first.columnLong);
         assertEquals(allTypes.columnBoolean, first.columnBoolean);
@@ -189,6 +190,7 @@ public class BulkInsertTests {
 
         NullTypes first = realm.where(NullTypes.class).findFirst();
 
+        assertNotNull(first);
         assertEquals(nullTypes1.getId(), first.getId());
         assertEquals(nullTypes1.getFieldIntegerNull(), first.getFieldIntegerNull());
         assertEquals(nullTypes1.getFieldFloatNull(), first.getFieldFloatNull());
@@ -226,6 +228,7 @@ public class BulkInsertTests {
 
         first = realm.where(NullTypes.class).equalTo("id", 1).findFirst();
 
+        assertNotNull(first);
         assertEquals(nullTypes1.getId(), first.getId());
         assertNull(first.getFieldIntegerNull());
         assertNull(first.getFieldFloatNull());
@@ -296,6 +299,7 @@ public class BulkInsertTests {
         realm.commitTransaction();
 
         primaryKeyAsString = realm.where(PrimaryKeyAsString.class).isNull("name").findFirst();
+        assertNotNull(primaryKeyAsString);
         assertNull(primaryKeyAsString.getName());
         assertEquals(19, primaryKeyAsString.getId());
 
@@ -307,6 +311,7 @@ public class BulkInsertTests {
         realm.commitTransaction();
 
         primaryKeyAsShort = realm.where(PrimaryKeyAsBoxedShort.class).isNull("id").findFirst();
+        assertNotNull(primaryKeyAsShort);
         assertNull(primaryKeyAsShort.getId());
         assertEquals("42", primaryKeyAsShort.getName());
     }
@@ -348,6 +353,7 @@ public class BulkInsertTests {
         AllTypesPrimaryKey obj = realm.where(AllTypesPrimaryKey.class).findFirst();
 
         // Check that the only element has all its properties updated
+        assertNotNull(obj);
         assertEquals("Bar", obj.getColumnString());
         assertEquals(1, obj.getColumnLong());
         assertEquals(2.23F, obj.getColumnFloat(), 0);
@@ -474,7 +480,9 @@ public class BulkInsertTests {
         });
 
         assertEquals(1, realm.where(PrimaryKeyAsLong.class).count());
-        assertEquals("Baz", realm.where(PrimaryKeyAsLong.class).findFirst().getName());
+        PrimaryKeyAsLong first = realm.where(PrimaryKeyAsLong.class).findFirst();
+        assertNotNull(first);
+        assertEquals("Baz", first.getName());
     }
 
     @Test
@@ -679,6 +687,7 @@ public class BulkInsertTests {
 
         realm.beginTransaction();
         try {
+            //noinspection ConstantConditions
             realm.insert(nullObject);
             fail("Should trigger NullPointerException");
         } catch (IllegalArgumentException ignore) {
@@ -694,6 +703,7 @@ public class BulkInsertTests {
 
         realm.beginTransaction();
         try {
+            //noinspection ConstantConditions
             realm.insert(nullObjects);
             fail("Should trigger IllegalArgumentException");
         } catch (IllegalArgumentException ignore) {
@@ -776,5 +786,25 @@ public class BulkInsertTests {
         assertNotNull(first.getFieldObject());
         assertEquals(42, first.getFieldObject().getFieldLong());
         assertEquals(2, realm.where(AllJavaTypes.class).count());
+    }
+
+    @Test(expected = IllegalStateException.class)
+    public void insert_collection_notInTransaction() {
+        realm.insert(Arrays.asList(new AllTypes(), new AllTypes()));
+    }
+
+    @Test(expected = IllegalStateException.class)
+    public void insert_object_notInTransaction() {
+        realm.insert(new AllTypes());
+    }
+
+    @Test(expected = IllegalStateException.class)
+    public void insertOrUpdate_collection_notInTransaction() {
+        realm.insert(Arrays.asList(new AllTypesPrimaryKey(), new AllTypesPrimaryKey()));
+    }
+
+    @Test(expected = IllegalStateException.class)
+    public void insertOrUpdate_object_notInTransaction() {
+        realm.insert(new AllTypes());
     }
 }